+++ server/channelserver/handlers_cast_binary.go
@@ -201,13 +201,13 @@ func parseChatCommand(s *Session, command string) {
 				name = strings.ToLower(name)
 				for _, course := range mhfcourse.Courses() {
 					for _, alias := range course.Aliases() {
-						if strings.ToLower(name) == strings.ToLower(alias) {
+						if strings.EqualFold(name, alias) {
 							if slices.Contains(s.server.erupeConfig.Courses, config.Course{Name: course.Aliases()[0], Enabled: true}) {
 								var delta, rightsInt uint32
 								if mhfcourse.CourseExists(course.ID, s.courses) {
 									ei := slices.IndexFunc(s.courses, func(c mhfcourse.Course) bool {
 										for _, alias := range c.Aliases() {
-											if strings.ToLower(name) == strings.ToLower(alias) {
+											if strings.EqualFold(name, alias) {
 												return true
 											}
 										}