+++ server/channelserver/sys_stage.go
@@ -68,7 +68,7 @@ func (s *Stage) BroadcastMHF(pkt mhfpacket.MHFPacket, ignoredSession *Session) {
 	s.Lock()
 	defer s.Unlock()
 	for session := range s.clients {
-		if session == ignoredSession {
+		if session == ignoredSession || session.clientContext == nil {
 			continue
 		}